Change `REFRESHES_PER_SESSION` from `u128` to `u32` (#513)
* Change refresh session type to `u32`

* Fix typo

* Remove `Result` return type

* Remove `unwrap`s from test

+++ crypto/server/src/helpers/substrate.rs
@@ -156,7 +156,7 @@ pub async fn get_key_visibility(
 pub async fn get_refreshes_done(
     api: &OnlineClient<EntropyConfig>,
     rpc: &LegacyRpcMethods<EntropyConfig>,
-) -> Result<u128, ProtocolErr> {
+) -> Result<u32, ProtocolErr> {
     let block_hash = rpc
         .chain_get_block_hash(None)
         .await?
